+++ src/RemoteStorageDetails/RemoteStorageDetails.js
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import { useIntl } from 'react-intl';
 import {
   useHistory,
   useParams,
@@ -39,6 +39,7 @@ const RemoteStorageDetails = ({
 }) => {
   const history = useHistory();
   const { id } = useParams();
+  const intl = useIntl();
 
   const [expandAll, sections, toggleSection] = useAccordionToggle(
     Object.values(SECTIONS_STORAGE).reduce((acc, k) => {
@@ -59,7 +60,7 @@ const RemoteStorageDetails = ({
         }}
       >
         <Icon size="small" icon="edit">
-          <FormattedMessage id="ui-remote-storage.edit" />
+          {intl.formatMessage({ id: 'ui-remote-storage.edit' })}
         </Icon>
       </Button>
 
@@ -70,11 +71,12 @@ const RemoteStorageDetails = ({
         onClick={onRemovestorage}
       >
         <Icon size="small" icon="trash">
-          <FormattedMessage id="ui-remote-storage.delete" />
+          {intl.formatMessage({ id: 'ui-remote-storage.delete' })}
         </Icon>
       </Button>
     </MenuSection>
-  ), [id, onRemovestorage]);
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  ), [id]);
 
   const closePane = useCallback(
     () => {
@@ -111,46 +113,53 @@ const RemoteStorageDetails = ({
         onToggle={toggleSection}
       >
         <Accordion
-          label={<FormattedMessage id="ui-remote-storage.details.title" />}
+          label={intl.formatMessage({ id: 'ui-remote-storage.details.title' })}
           id={SECTIONS_STORAGE.INFORMATION}
         >
           {storage.metadata && <ViewMetaData metadata={storage.metadata} />}
           <Row>
             <Col xs={6}>
-              <KeyValue label={<FormattedMessage id="ui-remote-storage.details.storageName" />}>
+              <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.details.storageName' })}>
                 {storage.name}
               </KeyValue>
             </Col>
             <Col xs={4}>
-              <KeyValue label={<FormattedMessage id="ui-remote-storage.details.providerName" />}>
-                {storage.providerName}
+              <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.details.providerName' })}>
+                {intl.formatMessage({ id: `ui-remote-storage.name.${storage.providerName}` })}
               </KeyValue>
             </Col>
             <Col xs={6}>
-              <KeyValue label={<FormattedMessage id="ui-remote-storage.details.url" />}>
+              <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.details.url' })}>
                 {storage.url}
               </KeyValue>
             </Col>
+            {storage.statusUrl && (
+              <Col xs={6}>
+                <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.details.statusUrl' })}>
+                  {storage.statusUrl}
+                </KeyValue>
+              </Col>
+            )}
           </Row>
         </Accordion>
 
         <Accordion
-          label={<FormattedMessage id="ui-remote-storage.accession.title" />}
+          label={intl.formatMessage({ id: 'ui-remote-storage.accession.title' })}
           id={SECTIONS_STORAGE.ACCESSION}
         >
-          <KeyValue label={<FormattedMessage id="ui-remote-storage.accession.schedule.title" />}>
-            <FormattedMessage
-              id="ui-remote-storage.accession.schedule.info"
-              values={{
+          <KeyValue label={intl.formatMessage({ id: 'ui-remote-storage.accession.schedule.title' })}>
+            {intl.formatMessage(
+              { id: 'ui-remote-storage.accession.schedule.info' },
+              {
                 delay: storage.accessionDelay,
                 unit: storage.accessionTimeUnit,
-              }}
-            />
+              },
+            )}
           </KeyValue>
         </Accordion>
 
         <Accordion
-          label={<FormattedMessage id="ui-remote-storage.retrieval.title" />}
+          label={intl.formatMessage({ id: 'ui-remote-storage.retrieval.title' })}
           id={SECTIONS_STORAGE.RETRIEVAL}
         />
       </AccordionSet>