+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 * [UIRS-3](https://issues.folio.org/browse/UIRS-3) Permission: Remote storage configuration
 * [UIRS-2](https://issues.folio.org/browse/UIRS-2) Remote storage configuration
 * [UIRS-6](https://issues.folio.org/browse/UIRS-6) Edit remote storage configuration
+* [UIRS-7](https://issues.folio.org/browse/UIRS-7) Remove/delete remote storage configuration